# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lounge Queue API Documentation",
        "description": "API documentation for the Lounge Queue system",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/visitor/lounge/queue": {
            "post": {
                "tags": [
                    "Lounge Queue"
                ],
                "summary": "Add visitor to queue",
                "description": "Add a visitor to the waiting room queue",
                "operationId": "enqueueVisitor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "visitor_id"
                                ],
                                "properties": {
                                    "visitor_id": {
                                        "description": "ID of the visitor to add to queue",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Visitor successfully added to queue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "position": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Successfully added to waiting room queue"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Visitor already in queue"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lounge Queue"
                ],
                "summary": "Remove visitor from queue",
                "description": "Remove a visitor from the waiting room queue",
                "operationId": "exitQueue",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "visitor_id"
                                ],
                                "properties": {
                                    "visitor_id": {
                                        "description": "ID of the visitor to remove from queue",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Visitor successfully removed from queue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully removed from queue"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Visitor not found in queue"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provider/lounge/list": {
            "get": {
                "tags": [
                    "Lounge Queue"
                ],
                "summary": "Get waiting list",
                "description": "Get the current waiting room queue list",
                "operationId": "getWaitingList",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved waiting list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "visitors": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "position": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "visitor_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "visitor_name": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            },
                                                            "reason": {
                                                                "type": "string",
                                                                "nullable": true
                                                            },
                                                            "waiting_time": {
                                                                "type": "string",
                                                                "example": "5 minutes ago"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/provider/lounge/pickup": {
            "post": {
                "tags": [
                    "Lounge Queue"
                ],
                "summary": "Pick up visitor",
                "description": "Pick up a visitor from the waiting room queue",
                "operationId": "pickupVisitor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "visitor_id"
                                ],
                                "properties": {
                                    "visitor_id": {
                                        "description": "ID of the visitor to pick up",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Visitor successfully picked up",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "visitor_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "visitor_name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "waited_time": {
                                                    "type": "string",
                                                    "example": "5 minutes ago"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Successfully picked up visitor from queue"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Visitor not found in queue"
                    },
                    "409": {
                        "description": "Provider is busy"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/lounge/queue": {
            "get": {
                "tags": [
                    "Lounge Queue"
                ],
                "summary": "Get current visitor's queue item",
                "description": "Get the current visitor's position and details in the lounge queue",
                "operationId": "getQueueItemByCurrentVisitor",
                "responses": {
                    "200": {
                        "description": "Queue item details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "position": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "joined_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-03-29T10:00:00Z"
                                                },
                                                "waited_time": {
                                                    "type": "string",
                                                    "example": "5 minutes"
                                                },
                                                "estimated_wait_time": {
                                                    "type": "string",
                                                    "example": "10 minutes"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Visitor not found in queue"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}